// src/types.ts
export type TMode =
  | 'Приветственные'
  | 'Прочие'
  | 'Все слова'
  | 'Вопросы'
  | 'Цифры (0-9)'
  | 'Цифры (10-90)'
  | 'Цифры'
  | 'Местоимения'
  | 'Прилагательные'
  | 'Прочие о человеке'
  | 'Профессии'
  | 'Семья и друзья'
  | 'Глаголы'